# web/src/main/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Revet Auth API
  description: OAuth 2.1/OIDC Server API
  termsOfService: https://revethq.com/terms
  contact:
    name: Revet
    url: https://revethq.com
    email: info@revethq.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: http://localhost:5000
  description: Auto generated value
- url: http://0.0.0.0:5000
  description: Auto generated value
...
paths:
...
  /users:
    get:
      tags:
      - Users
      summary: List users with optional filtering
      parameters:
      - name: authorizationServerIds
        in: query
        schema:
          type: array
          items:
            format: uuid
            pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
            type: string
          nullable: true
      - name: limit
        in: query
        schema:
          format: int32
          type: integer
          nullable: true
      - name: offset
        in: query
        schema:
          format: int32
          type: integer
          nullable: true
      responses:
        "200":
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
    post:
      tags:
      - Users
      summary: Create a new user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        "200":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "400":
          description: Invalid request
  /users/{userId}:
    get:
      tags:
      - Users
      summary: Get a user by ID
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "404":
          description: User not found
    put:
      tags:
      - Users
      summary: Update a user by ID
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "404":
          description: User not found
    delete:
      tags:
      - Users
      summary: Delete a user by ID
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: User deleted successfully
        "404":
          description: User not found
...
components:
  schemas:
...
    Metadata:
      type: object
      properties:
        properties:
          type: object
          additionalProperties: {}
          nullable: true
        identifiers:
          type: array
          items:
            $ref: '#/components/schemas/MetadataIdentifiersInner'
          nullable: true
        schemaValidations:
          type: array
          items:
            $ref: '#/components/schemas/MetadataSchemaValidationsInner'
          nullable: true
    MetadataIdentifiersInner:
      type: object
      properties:
        system:
          type: string
          nullable: true
        value:
          type: string
          nullable: true
    MetadataSchemaValidationsInner:
      type: object
      properties:
        schemaId:
          type: string
          allOf:
          - $ref: '#/components/schemas/UUID'
          nullable: true
        isValid:
          type: boolean
          nullable: true
        validatedOn:
          type: string
          allOf:
          - $ref: '#/components/schemas/OffsetDateTime'
          nullable: true
    OffsetDateTime:
      format: date-time
      type: string
      example: 2022-03-10T12:15:50-04:00
    Page:
      type: object
      properties:
        next:
          type: string
          nullable: true
        previous:
          type: string
          nullable: true
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
    UserRequest:
      type: object
      properties:
        authorizationServerId:
          type: string
          allOf:
          - $ref: '#/components/schemas/UUID'
          nullable: true
        username:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        profile: {}
        password:
          type: string
          nullable: true
        metadata:
          type: object
          allOf:
          - $ref: '#/components/schemas/Metadata'
          nullable: true
    UserResponse:
      type: object
      properties:
        id:
          type: string
          nullable: true
        authorizationServerId:
          type: string
          allOf:
          - $ref: '#/components/schemas/UUID'
          nullable: true
        username:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        profile: {}
        createdOn:
          type: string
          allOf:
          - $ref: '#/components/schemas/OffsetDateTime'
          nullable: true
        updatedOn:
          type: string
          allOf:
          - $ref: '#/components/schemas/OffsetDateTime'
          nullable: true
        metadata:
          type: object
          allOf:
          - $ref: '#/components/schemas/Metadata'
          nullable: true
    UsersResponse:
      required:
      - page
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
          nullable: true
        page:
          type: object
          allOf:
          - $ref: '#/components/schemas/Page'
          nullable: true
...